refactored ReviewsBox into it's own folder/component and off the home page then added links to Ratings and Reviews to the Nav bar, moved ratings below reviews on the home page
